cli: color nodes by branch in log graph
Colors each node in the log graph based on branch, found by searching from each
head and coloring the parents.

The coloring is unstable and does not map a specific branch or change ID to a
specific color. The colors are picked in order of branch head appearance for the
specific log command being run.

Cannot piggy-back on `item` in graph.rs as it does not yet exist for the
parents, when the parents are discovered by a child node.

# The journey of a graph node

RevSet::iter_graph() produces nodes with parent edges. These are fed to
TopoGroupedGraphIterator, which produces nodes with child edges, needed because
children are drawn before parents in the graph output.

Finally, the nodes are given to GraphLog::add_node() one by one. GraphLog
renders each node using SaplingGraphLog::next_row(), calling e.g.
BoxDrawingRenderer::next_row().

The coloring of the nodes happen in log::cmd_log(), by modifying the node symbol
print.